AWS-lambda-implementation/requirements.txt
beautifulsoup4==4.7.1
requests==2.21.0
boto==2.49.0
urllib3==1.24.1

